Adds `GLM_EXT_matrix_elementwise_comparison`
The GLSL 4.20.8 specification provides relational operators/functions
for vector types in "Section 8.7 : Vector Relational Functions".

This extension provides the same element-wise comparison functors for
matrix types. The return type is a boolean matrix of the same dimensions
as the inputs, representing the result of the element-wise comparisons.

In order not to create any name collisions with the existing vector-wise
comparison operators (in `GLM_EXT_matrix_relational`), the functors
added in this patch are declared in the `glm::elementwise` namespace.

New functions added :
  - `mat<bool> glm::elementwise::lessThan(mat<T>&, mat<T>&)`
  - `mat<bool> glm::elementwise::lessThanEqual(mat<T>&, mat<T>&)`
  - `mat<bool> glm::elementwise::greaterThan(mat<T>&, mat<T>&)`
  - `mat<bool> glm::elementwise::greaterThanEqual(mat<T>&, mat<T>&)`
  - `mat<bool> glm::elementwise::equal(mat<T>&, mat<T>&)`
  - `mat<bool> glm::elementwise::notEqual(mat<T>&, mat<T>&)`