Adds matrix support to 'GLM_GTC_epsilon'
This commit fully enables the use of the `GLM_GTC_epsilon` extension to
the matrix types.

It is now possible to check the element-wise equality or inequality
(with a epsilon-tolerance) of matrices, just as was previously possible
with vector types.

This commit also adds the necessary tests for the `GLM_GTC_epsilon`
extension in `test/gtc/gtc_epsilon.cpp`.